#42 zipinfo zip64 (#50)

+++ src/main/java/ru/olegcherednik/zip4jvm/model/builders/EndCentralDirectoryBuilder.java
@@ -27,7 +27,7 @@ public final class EndCentralDirectoryBuilder {
     }
 
     private int getTotalDisks() {
-        return zipModel.isZip64() ? ZipModel.MAX_TOTAL_DISKS : (int)zipModel.getTotalDisks();
+        return zipModel.isZip64() ? ZipModel.MAX_TOTAL_DISKS : zipModel.getTotalDisks();
     }
 
     private int getDiskEntries() {
